# aibots-api-develop/openapi/full.yaml
openapi: 3.0.3
info:
  title: AIBots Full API
  description: This document describes the Full API interface for AIBots
  version: 1.0.0
  contact:
    name: Moonshot Team
    url: 'https://www.tech.gov.sg/capability-centre-dsaid'
    email: david_tw_lee@tech.gov.sg
servers:
  - url: 'https://aibots.gov.sg/v{major}.{minor}'
    description: Full server address hosted on cloud
    variables:
      major:
        enum:
          - "1"
        default: "1"
      minor:
        enum:
          - "0"
        default: "0"
...
paths:
...
  /definitions/agents:
    get:
      tags:
        - Schemas
      summary: Default settings form parameters for AI Agents
      description: >-
        Retrieves the default form parameters available for configuring AI 
        Agents
      operationId: get_agent_definitions
      responses:
        '200':
          description: Successfully retrieved Agent settings
          content:
            application/json:
              schema:
                description: Default Agent settings definitions
                type: array
                items:
                  $ref: '#/components/schemas/AtlasParam'
                default: []
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/PermissionsError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Agent APIs
  /agents:
    post:
      tags:
        - Agents
      summary: Creates a new Agent
      description: |-
        An Agent encapsulates various forms of interacting with a Large Language 
        Model (LLM), this could be as diverse as interacting with an image 
        generation model (multi-modal Chat Agent), to performing question and 
        answer on a specified knowledge base (RAG Agent), to performing actions
        on a user's behalf (AI Agent). An Agent can then be served out for other 
        users to interact with it. Agent are highly configurable, this provides 
        Agent creators with the flexibility to serve out Agents that address the 
        needs of their target audience. Agent access can be restricted to 
        different individuals or organisations/groups. They can be cloned with 
        the entire Agent configuration and customised according to their needs. 
        This API call creates a new Agent configuration that structures the 
        interaction.
      operationId: create_agent
      requestBody:
        description: Configuration of Agent to be created
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentPOST'
      responses:
        '201':
          $ref: '#/components/responses/CreatedResponse'
        '400':
          $ref: '#/components/responses/InvalidParametersError'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/PermissionsError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      tags:
        - Agents
      summary: Retrieves all Agents visible to the user
      description: >-
        Returns a list of all Agents that the user is permitted to view and 
        invoke. Supports various query parameters for retrieving public and 
        shared Agents.
      operationId: get_agents
      parameters:
        - $ref: '#/components/parameters/IDListParam'
        - name: featured
          description: Retrive all featured Agents
          in: query
          required: false
          schema:
            type: boolean
        - name: public
          description: Retrieve all Agents available to the WOG
          in: query
          required: false
          schema:
            type: boolean
        - name: owner
          description: Retrieve all Agents that the user owns
          in: query
          required: false
          schema:
            type: boolean
        - name: admin
          description: Retrieve all Agents that the user can edit/manage
          in: query
          required: false
          schema:
            type: boolean
        - name: shared
          description: Retrieve all Agents shared with the user
          in: query
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Successfully retrieved Agents
          content:
            application/json:
              schema:
                description: List of Agents
                type: array
                items:
                  $ref: '#/components/schemas/Agent'
                default: []
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/PermissionsError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  '/agents/{id}':
    parameters:
      - $ref: '#/components/parameters/IDParam'
    put:
      tags:
        - Agents
      summary: Updates the core configuration of an existing Agent
      description: |-
        This API enables the user to update the main configuration of an 
        existing Agent.
      operationId: update_agent
      requestBody:
        description: Configuration of Agent to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentPUT'
      responses:
        '204':
          $ref: '#/components/responses/UpdatedResponse'
        '400':
          $ref: '#/components/responses/InvalidParametersError'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/PermissionsError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      tags:
        - Agents
      summary: Retrieves the configuration of an existing Agent
      description: Retrieves an existing Agent
      operationId: get_agent
      responses:
        '200':
          description: Successfully retrieved Agent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/InvalidParametersError'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/PermissionsError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Agents
      summary: Deletes an existing Agent configuration
      description: |-
        Includes the knowledgeBase boolean flag to delete the Knowledge Base 
        associated with the Agent. If the Knowledge Base is shared by other Agents,
        then the files boolean flag will not apply. A push notification will be
        sent upon completion.
      operationId: delete_agent
      parameters:
        - description: Deletes all files associated with the Knowledge Bases
          name: files
          in: query
          required: false
          schema:
            type: boolean
      responses:
        '202':
          description: |-
            Successfully deleted Agent and associated Knowledge Base will be 
            cleared as a background task with a push notification sent.
        '204':
          $ref: '#/components/responses/DeletedResponse'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/PermissionsError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/InvalidDeletionError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  '/agents/approvals':
    get:
      tags:
        - Agents
      summary: Retrieve all approval details for processing
      description: >-
        This is administrative feature for superusers to view all approvals
        that have been created. Supports querying by ID via the ids query 
        parameter.
      operationId: get_approvals
      parameters:
        - $ref: '#/components/parameters/IDListParam'
        - description: Indicates that all pending Agent Approvals will be returned
          name: pending
          in: query
          required: false
          schema:
            type: boolean
        - description: Indicates that approved Agents will be returned
          name: approved
          in: query
          required: false
          schema:
            type: boolean
        - description: Indicates that rejected Agents will be returned
          name: rejected
          in: query
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Successfully retrieved all approvals
          content:
            application/json:
              schema:
                description: List of all approvals
                type: array
                items:
                  $ref: '#/components/schemas/AgentApproval'
                default: []
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/PermissionsError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  '/agents/approvals/{id}':
    parameters:
      - $ref: '#/components/parameters/IDParam'
    post:
      tags:
        - Agents
      summary: Submit approval details for processing
      description: >-
        Before the Agent is able to be published, it needs to be approved. This 
        set of endpoints models the Approval workflow, wherein an 
        email/notification is sent to the approver. Thereafter the Approver 
        must make a PUT API call to the /agents/approvals/{id} endpoint to 
        approve the request.
      operationId: create_approval
      responses:
        '204':
          $ref: '#/components/responses/UpdatedResponse'
        '400':
          $ref: '#/components/responses/InvalidParametersError'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/PermissionsError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  '/agents/approvals/{id}/approves':
    parameters:
      - $ref: '#/components/parameters/IDParam'
    post:
      tags:
        - Agents
      summary: Approve Agent for release
      description: >-
        Only superusers or the approver specified is able to make a call 
        to this endpoint. After an Agent is in production, only then can they 
        set the visibility status of the Agent to wog or public.
      operationId: approve_approval
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentApprovalPUT"
      responses:
        '204':
          $ref: '#/components/responses/UpdatedResponse'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/PermissionsError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  '/agents/approvals/{id}/rejects':
    parameters:
      - $ref: '#/components/parameters/IDParam'
    post:
      tags:
        - Agents
      summary: Rejects an Agent release Approval
      description: >-
        Only superusers or the approver specified is able to make a call 
        to this endpoint. If the Agent is already in production status, then
        we will need to revoke the production status and set the visibility
        status of the Agent back to private.
      operationId: reject_approval
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentApprovalPUT"
      responses:
        '204':
          $ref: '#/components/responses/UpdatedResponse'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/PermissionsError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # File Uploads
  # Functionality provided by atlas-archeía
  # TODO: API to manage folders and download multiple files
...
security:
  - AtlasAuth: []
components:
  schemas:
...
    AtlasParam:
      description: Representation of a dynamic configurable parameter field
      type: object
      properties:
        name:
          description: Display name of the Parameter
          type: string
        description:
          description: Brief description of the Parameter
          type: string
          default: ''
        key:
          description: Internal key of the Parameter
          type: string
        default:
          description: Default parameter value
          nullable: true
          default: null
        required:
          description: Indicates if the Parameter is required
          type: boolean
          default: false
        schema:
          description: Json schema structure to validate the Parameter against
          type: object
          additionalProperties: true
          default: {}
        tags:
          description: Tags associated with the Parameter
          type: array
          items:
            type: string
          default: []
        componentConfig:
          description: UI components to render the Param configuration form with
          anyOf:
            - $ref: '#/components/schemas/AtlasComponents'
            - description: List of Atlas Components
              type: array
              items:
                $ref: '#/components/schemas/AtlasComponents'
          nullable: true
          default: null
      required:
        - name
        - description
        - key
        - default
        - schema
        - required
        - tags
        - componentConfig

    # States
...
    AgentPOST:
      description: POST request body for creating a Agent
      type: object
      properties:
        name:
          description: Name of the Agent
          type: string
        description:
          description: Brief description of the Agent
          type: string
        ownership:
          $ref: '#/components/schemas/Ownership'
        files:
          description: Non-embedded files associated with the Agent
          type: array
          items:
            type: string
            format: uuid
          default: []
        welcomeMessage:
          description: Welcome message the Agent will display when a new chat is created
          type: string
          default: "Hello! I am your assistant. How can I help you?"
        agency:
          description: Agency associated with the Agent
          type: string
        templates:
          description: Default templates to use with the Agent
          type: array
          items:
            $ref: '#/components/schemas/AgentTemplate'
          default: []
        knowledgeBases:
          description: Knowledge bases associated with Agent
          type: array
          items:
            $ref: '#/components/schemas/DataSource'
          default: []
        defaultPipeline:
          description: Default RAG pipeline to be used
          type: string
          nullable: true
          default: null
        rags:
          description: List of supported RAG pipelines
          type: array
          items:
            $ref: '#/components/schemas/RAGConfig'
          default: []
        chat:
          $ref: '#/components/schemas/AgentChatConfig'
        tools:
          description: Tools that the Agent is able to access
          type: array
          items:
            type: string
            format: uuid
          default: []
        settings:
          description: >-
            Additional settings associated with the Agent
          type: object
          additionalProperties: true
          default: {}
          example:
            conversationHistory: true
            accessDocs: true
            citations: true
            userCreativity: true
            userMaxLength: true
            userRemovePII: true
            userCustomInstructions: true
      required:
        - name
        - agency
...
  parameters:
...
    IDListParam:
      name: ids
      description: A list of optional IDs
      in: query
      schema:
        type: array
        items:
          type: string
          format: uuid
      style: form
      explode: true
...
  responses:
    CreatedResponse:
      description: Successfully created resource with ID
      content:
        application/json:
          schema:
            description: Generated resource ID
            type: object
            properties:
              id:
                description: Unique ID of the resource
                type: string
                format: uuid
            required:
              - id
    UpdatedResponse:
      description: Successfully updated specified resource
...
    InvalidParametersError:
      description: Invalid parameter(s)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
          example:
            code: 400
            message: Invalid parameters provided
    AuthenticationError:
      description: Authentication error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
          example:
            code: 401
            message: Unauthorised usage of the API
    PermissionsError:
      description: Insufficient permissions error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
          example:
            code: 403
            message: Insufficient permissions to use API
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
          example:
            code: 500
            message: Internal server error

    # SCIM responses
...
  securitySchemes:
    AtlasAuth:
      description: Atlas JWT authentication token
      type: http
      scheme: bearer
      bearerFormat: JWT
    AtlasAPIKey:
      description: Atlas API Key authentication
      type: apiKey
      in: header
      name: X-ATLAS-Key